Geometric factors: Move structs to separate GeomData module to avoid circular dependency issue
