+++ src/Numerics/Mesh/Mesh.jl
@@ -1,6 +1,7 @@
 module Mesh
 include("BrickMesh.jl")
 include("Topologies.jl")
+include("GeomData.jl")
 include("Metrics.jl")
 include("Elements.jl")
 include("Grids.jl")
